Added solution - LeetHub

+++ Friends Pairing Problem - GFG/friends-pairing-problem.cpp
@@ -8,11 +8,11 @@ class Solution
 public:
     int countFriendsPairings(int n) { 
         int mod=1e9+7;
-        long long int a = 1, b = 2, c = 0;
+        int a = 1, b = 2, c = 0;
         if (n <= 2) {
             return n;
         }
-        for (int i = 3; i <= n; i++) {
+        for (long long int i = 3; i <= n; i++) {
             c=(b+(i-1)*a)%mod;
             a=b;
             b=c;
